@@ -1,6 +1,6 @@
 {
   "name": "expenses-tracker",
-  "version": "1.0.0",
+  "version": "2.6.1-SNAPSHOT",
   "license": "MIT",
   "scripts": {
     "ng": "nx",
@@ -14,64 +14,74 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "16.0.6",
-    "@angular/cdk": "16.0.4",
-    "@angular/common": "16.0.6",
-    "@angular/compiler": "16.0.6",
-    "@angular/core": "16.0.6",
+    "@angular/animations": "16.2.1",
+    "@angular/cdk": "16.2.1",
+    "@angular/common": "16.2.1",
+    "@angular/compiler": "16.2.1",
+    "@angular/core": "16.2.1",
     "@angular/fire": "^7.5.0",
-    "@angular/forms": "16.0.6",
-    "@angular/material": "16.0.4",
-    "@angular/platform-browser": "16.0.6",
-    "@angular/platform-browser-dynamic": "16.0.6",
-    "@angular/router": "16.0.6",
-    "firebase": "^9.22.2",
+    "@angular/forms": "16.2.1",
+    "@angular/platform-browser": "16.2.1",
+    "@angular/platform-browser-dynamic": "16.2.1",
+    "@angular/router": "16.2.1",
+    "@nx/angular": "16.7.2",
+    "firebase": "^10.2.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "uuid": "^9.0.0",
-    "zone.js": "0.13.1",
-    "@nx/angular": "16.3.2"
+    "zone.js": "0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "16.0.6",
-    "@angular-devkit/core": "16.0.6",
-    "@angular-devkit/schematics": "16.0.6",
+    "@angular-devkit/build-angular": "16.2.0",
+    "@angular-devkit/core": "16.2.0",
+    "@angular-devkit/schematics": "16.2.0",
     "@angular-eslint/eslint-plugin": "16.0.3",
     "@angular-eslint/eslint-plugin-template": "16.0.3",
     "@angular-eslint/template-parser": "16.0.3",
-    "@angular/cli": "~16.0.0",
-    "@angular/compiler-cli": "16.0.6",
-    "@angular/language-service": "16.0.6",
-    "@schematics/angular": "16.0.6",
+    "@angular/cli": "~16.2.0",
+    "@angular/compiler-cli": "16.2.1",
+    "@angular/language-service": "16.2.1",
+    "@nx/cypress": "16.7.2",
+    "@nx/eslint-plugin": "16.7.2",
+    "@nx/jest": "16.7.2",
+    "@nx/js": "16.7.2",
+    "@nx/linter": "16.7.2",
+    "@nx/storybook": "16.7.2",
+    "@nx/web": "16.7.2",
+    "@nx/workspace": "16.7.2",
+    "@schematics/angular": "16.2.0",
+    "@storybook/addon-essentials": "^7.2.1",
+    "@storybook/addon-interactions": "^7.2.1",
+    "@storybook/angular": "^7.2.1",
+    "@storybook/core-server": "^7.2.1",
+    "@storybook/jest": "~0.1.0",
+    "@storybook/test-runner": "^0.13.0",
+    "@storybook/testing-library": "~0.2.0",
     "@types/jest": "^29.4.0",
     "@types/node": "16.11.7",
     "@types/uuid": "^9.0.1",
-    "@typescript-eslint/eslint-plugin": "5.60.0",
-    "@typescript-eslint/parser": "5.60.0",
+    "@typescript-eslint/eslint-plugin": "5.62.0",
+    "@typescript-eslint/parser": "5.62.0",
     "autoprefixer": "^10.4.0",
-    "cypress": "12.15.0",
-    "eslint": "~8.15.0",
-    "eslint-config-prettier": "8.1.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-prettier": "^4.2.1",
+    "cypress": "12.17.4",
+    "eslint": "8.46.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-cypress": "2.14.0",
+    "eslint-plugin-prettier": "^5.0.0",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
     "jest-preset-angular": "13.1.1",
-    "nx": "16.3.2",
+    "nx": "16.7.2",
+    "nx-cloud": "16.4.0-beta.1",
     "postcss": "^8.4.5",
-    "prettier": "^2.6.2",
+    "prettier": "^3.0.2",
     "source-map-explorer": "^2.5.3",
-    "tailwindcss": "^3.0.2",
+    "tailwindcss": "^3.3.3",
     "ts-jest": "29.1.0",
     "ts-node": "10.9.1",
-    "typescript": "5.0.4",
-    "nx-cloud": "16.0.5",
-    "@nx/workspace": "16.3.2",
-    "@nx/cypress": "16.3.2",
-    "@nx/js": "16.3.2",
-    "@nx/linter": "16.3.2",
-    "@nx/eslint-plugin": "16.3.2",
-    "@nx/jest": "16.3.2",
-    "@nx/web": "16.3.2"
+    "typescript": "5.1.6"
   }
 }
+
