@@ -1,41 +1,58 @@
+import { DialogModule } from '@angular/cdk/dialog';
 import { CommonModule } from '@angular/common';
-import { Component, OnInit } from '@angular/core';
-import { MatIconModule } from '@angular/material/icon';
-import { MatTooltipModule } from '@angular/material/tooltip';
+import { Component, OnDestroy, Signal, computed, inject } from '@angular/core';
+import { toObservable } from '@angular/core/rxjs-interop';
 import { RouterModule } from '@angular/router';
 import { User } from 'firebase/auth';
-import { Observable } from 'rxjs';
+import { Subscription, switchMap } from 'rxjs';
 
-import { ExtractInitialsPipe } from '@expenses-tracker/features/profile';
-import { AddPocketbookGraphicComponent } from '@expenses-tracker/shared/assets';
-import { IPocketbook } from '@expenses-tracker/shared/interfaces';
+import { AvatarComponent, ExtractInitialsPipe, TooltipModule } from '@expenses-tracker/shared/common';
 
 import { PocketbookDetailService } from './pocketbook-detail.service';
 
 @Component({
   selector: 'expenses-tracker-pocketbook-detail',
   standalone: true,
-  imports: [
-    CommonModule,
-    MatTooltipModule,
-    MatIconModule,
-    RouterModule,
-
-    ExtractInitialsPipe,
-    AddPocketbookGraphicComponent
-  ],
+  imports: [AvatarComponent, CommonModule, DialogModule, ExtractInitialsPipe, RouterModule, TooltipModule],
   templateUrl: './pocketbook-detail.component.html'
 })
-export class PocketbookDetailComponent implements OnInit {
-  pocketbook$!: Observable<IPocketbook | null>;
-  collaboratorList$!: Observable<User[]>;
-  owner$!: Observable<User | null>;
-  constructor(private _service: PocketbookDetailService) {}
+export class PocketbookDetailComponent implements OnDestroy {
+  #service = inject(PocketbookDetailService);
+  pocketbook = computed(() => this.#service.pocketbook());
+  collaboratorList = computed(() => this.#service.collaboratorList());
+  owner = computed(() => this.#service.owner()) as Signal<User>;
+  viewMode = computed(() => this.#service.viewMode());
+  flags = computed(() => this.#service.flags());
+  #pocketbookContributors$!: Subscription;
+  #transactionAddDialogSubscription$!: Subscription;
+  #recalculateBalanceSubscription$!: Subscription;
 
-  ngOnInit() {
-    this._service.initializeComponent();
-    this.pocketbook$ = this._service.watchPocketbook$();
-    this.collaboratorList$ = this._service.watchCollaboratorList$();
-    this.owner$ = this._service.watchOwner$();
+  constructor() {
+    // NOTE: @usamazansari: be very careful while using toObservable as it may cause memory leak
+    this.#pocketbookContributors$ = toObservable(this.pocketbook)
+      .pipe(switchMap(() => this.#service.watchPocketbookContributors$()))
+      .subscribe();
+  }
+
+  collaboratorListTrack(index: number, user: User) {
+    return user.uid;
+  }
+
+  recalculateBalance() {
+    this.#recalculateBalanceSubscription$ = this.#service.recalculateBalance().subscribe();
+  }
+
+  gotoTransactionList() {
+    this.#service.gotoTransactionList();
+  }
+
+  gotoSettings() {
+    this.#service.gotoSettings();
+  }
+
+  ngOnDestroy() {
+    this.#pocketbookContributors$?.unsubscribe();
+    this.#transactionAddDialogSubscription$?.unsubscribe();
+    this.#recalculateBalanceSubscription$?.unsubscribe();
   }
 }
