@@ -1,16 +1,11 @@
 import { provideImgixLoader } from '@angular/common';
 import { ApplicationConfig, importProvidersFrom, isDevMode } from '@angular/core';
 import { AngularFireModule } from '@angular/fire/compat';
-import {
-  USE_EMULATOR as AUTH_EMULATOR,
-  AngularFireAuthModule
-} from '@angular/fire/compat/auth';
-import {
-  AngularFirestoreModule,
-  USE_EMULATOR as FIRESTORE_EMULATOR
-} from '@angular/fire/compat/firestore';
+import { USE_EMULATOR as AUTH_EMULATOR, AngularFireAuthModule } from '@angular/fire/compat/auth';
+import { AngularFirestoreModule, USE_EMULATOR as FIRESTORE_EMULATOR } from '@angular/fire/compat/firestore';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { provideRouter, withEnabledBlockingInitialNavigation } from '@angular/router';
+
 import { appRoutes } from './app.routes';
 
 const devEnv = isDevMode();
@@ -27,9 +22,7 @@ const firebaseConfig = {
 export const appConfig: ApplicationConfig = {
   providers: [
     provideRouter(appRoutes, withEnabledBlockingInitialNavigation()),
-    provideImgixLoader(
-      devEnv ? 'http://localhost:8080' : 'https://ua-expenses-tracker.firebaseapp.com/'
-    ),
+    provideImgixLoader(devEnv ? 'http://localhost:8080' : 'https://ua-expenses-tracker.firebaseapp.com/'),
     importProvidersFrom(
       AngularFireModule.initializeApp(firebaseConfig),
       AngularFireAuthModule,
