@@ -1,6 +1,6 @@
-import { Injectable, inject } from '@angular/core';
+import { Injectable, computed, inject, signal } from '@angular/core';
 import { Router } from '@angular/router';
-import { BehaviorSubject, catchError, distinctUntilChanged, of, switchMap, tap } from 'rxjs';
+import { catchError, map, switchMap, tap, throwError } from 'rxjs';
 
 import { ContextService, FirestoreService } from '@expenses-tracker/core';
 import { NotificationService, RoutePaths } from '@expenses-tracker/shared/common';
@@ -14,88 +14,58 @@ export type ComponentFlags = {
   providedIn: 'root'
 })
 export class TransactionAddService {
-  #flags$ = new BehaviorSubject<ComponentFlags>({ addTransaction: INITIAL_FLAGS });
-  #flags: ComponentFlags = { addTransaction: INITIAL_FLAGS };
-
   #context = inject(ContextService);
   #firestore = inject(FirestoreService);
   #notification = inject(NotificationService);
   #router = inject(Router);
-
-  #setFlags(flags: ComponentFlags) {
-    this.#flags = flags;
-    this.#flags$.next(this.#flags);
-  }
+  pocketbook = computed(() => this.#context.pocketbook());
+  flags = signal<ComponentFlags>({ addTransaction: INITIAL_FLAGS });
 
   resetFlags() {
-    this.#setFlags({ addTransaction: INITIAL_FLAGS });
-  }
-
-  watchFlags$() {
-    return this.#flags$.asObservable();
+    this.flags.set({ addTransaction: INITIAL_FLAGS });
   }
 
   addTransaction$(transaction: ITransaction) {
     this.resetFlags();
-    this.#setFlags({
-      ...this.#flags,
-      addTransaction: {
-        ...this.#flags.addTransaction,
-        loading: true
-      }
-    });
+    this.flags.update(value => ({ ...value, addTransaction: { ...value.addTransaction, loading: true } }));
 
     return this.#firestore.createTransaction$(transaction).pipe(
       switchMap(response =>
-        this.#context.watchPocketbook$().pipe(
-          distinctUntilChanged(
-            previous => !(previous?.transactionList ?? []).includes(transaction.id)
-          ),
-          switchMap(pocketbook =>
-            this.#firestore.updatePocketbook$({
-              ...pocketbook,
-              transactionList: [...(pocketbook?.transactionList ?? []), response?.id ?? ''],
-              balance: this.#context.addTransactionCalculateBalance(transaction)
-            })
-          )
-        )
+        this.#firestore.updatePocketbook$({
+          ...this.pocketbook(),
+          transactionList: [...(this.pocketbook()?.transactionList ?? []), response?.id ?? ''],
+          balance: this.#context.addTransactionCalculateBalance(transaction)
+        })
       ),
+      map(pb => pb),
       tap(() => {
         this.#notification.success({
           title: 'Transaction added',
           description: 'The transaction has been added successfully'
         });
-        this.resetFlags();
-        this.#router.navigate([
-          RoutePaths.Pocketbook,
-          this.#context.getPocketbook()?.id,
-          RoutePaths.Transaction,
-          RoutePaths.EntityList
-        ]);
+        this.flags.update(value => ({
+          ...value,
+          addTransaction: { ...value.addTransaction, loading: false, success: true, fail: false }
+        }));
       }),
       catchError(error => {
         this.#notification.error({
           description: `${error}.`,
           title: 'Error adding transaction'
         });
-        this.#setFlags({
-          ...this.#flags,
-          addTransaction: {
-            ...this.#flags.addTransaction,
-            loading: false,
-            fail: true,
-            visible: true
-          }
-        });
-        return of(error);
+        this.flags.update(value => ({
+          ...value,
+          addTransaction: { ...value.addTransaction, loading: false, success: false, fail: true }
+        }));
+        return throwError(() => new Error(error));
       })
     );
   }
 
-  cancelAddTransaction() {
+  gotoTransactionList() {
     this.#router.navigate([
       RoutePaths.Pocketbook,
-      this.#context.getPocketbook()?.id,
+      this.pocketbook()?.id,
       RoutePaths.Transaction,
       RoutePaths.EntityList
     ]);
