@@ -1,48 +1,34 @@
 import { CommonModule } from '@angular/common';
-import { Component, Input, inject } from '@angular/core';
-import { MatDialog, MatDialogModule } from '@angular/material/dialog';
-import { MatIconModule } from '@angular/material/icon';
-import { BehaviorSubject, EMPTY, of, switchMap } from 'rxjs';
+import { Component, Input, inject, signal } from '@angular/core';
+import { EMPTY } from 'rxjs';
 
 import { ITransaction } from '@expenses-tracker/shared/interfaces';
+import { TooltipModule } from '@expenses-tracker/shared/common';
 
 import { TransactionListItemService } from './transaction-list-item.service';
-import { TransactionDeleteDialogComponent } from './transaction-delete-dialog.component';
+import { CategoryFormatterPipePipe } from '../../../pipes';
 
 @Component({
   selector: 'expenses-tracker-transaction-list-item',
   standalone: true,
-  imports: [CommonModule, MatIconModule, MatDialogModule],
+  imports: [CategoryFormatterPipePipe, CommonModule, TooltipModule],
   templateUrl: './transaction-list-item.component.html',
   styles: []
 })
 export class TransactionListItemComponent {
   #service = inject(TransactionListItemService);
-  #dialog = inject(MatDialog);
-  #transaction$ = new BehaviorSubject<ITransaction | null>(null);
-  @Input() set transaction(value: ITransaction | null) {
-    this.#transaction$.next(value);
-  }
-  get transaction() {
-    return this.#transaction$.getValue();
+  transaction = signal<ITransaction | null>(null);
+  @Input() set transactionInput(value: ITransaction | null) {
+    this.transaction.set(value);
   }
 
   gotoEditTransaction() {
-    this.#service.gotoEditTransaction(this.transaction as ITransaction);
+    this.#service.gotoEditTransaction(this.transaction() as ITransaction);
   }
 
   deleteTransaction() {
-    const _ref = this.#dialog.open(TransactionDeleteDialogComponent);
-
-    _ref
-      .afterClosed()
-      .pipe(
-        switchMap(result =>
-          result
-            ? this.#service.deleteTransaction$(this.transaction as ITransaction)
-            : of(EMPTY)
-        )
-      )
-      .subscribe();
+    const result = confirm('Delete Transaction?');
+    const deleteStream = result ? this.#service.deleteTransaction$(this.transaction() as ITransaction) : EMPTY;
+    // deleteStream.subscribe();
   }
 }
