@@ -1,2 +1,6 @@
-export * from './notification/notification.component';
-export * from './notification/notification.service';
+export * from './avatar';
+export * from './datepicker';
+export * from './notification';
+export * from './select';
+export * from './segmented-control';
+export * from './tooltip';
